Missing attribute test passing
- Fix test with missing attributes. Test that you’re still on the new
form and that errors are displayed.
- Modify form to display errors

If you encounter errors upon saving a record, you’ll get a response
back from the server that looks like { errors: { name: [“can’t be
blank”], description: [“can’t be blank”]} and so forth.  Those errors
get set on the model you’re saving and can be accessed in the template
using model.errors.