Mark task complete/incomplete with fontawesome checkboxes
Instead of using checkboxes, I’m attaching a checkbox icon next to the
button.  However:
(1) the corresponding action is not fired upon clicking the checkbox;
(2) the buttons look bad and are not aligned with the checkboxes; and
(3) this just seems kind of hacky.
Is there a better way I’m not seeing?